+++ quartz-js/components/tag/docs/guide.html.js
@@ -1,0 +1,34 @@
+Q.components.guide.js.ui.tag = {};
+Q.components.guide.js.ui.tag.container = {
+  view: function(ctrl) {
+    return m('section#guide--component-tag', [
+      m.component(Q.components.shared.intro.ui.container, {
+        title: 'Tags',
+        intro: 'Tags or applied filters, having the appearance of a button but give you a hover state with a remove action.'
+      }),
+      m.component(Q.components.shared.block.ui.container, {
+        title: 'Example',
+        component: {
+          view: function() {
+            return m('div', [
+              m.component(vhxm.components.shared.tag.ui.container, {
+                onShow: function() {
+                  window.alert('show something');
+                },
+                onRemove: function() {
+                  window.alert('remove something');
+                }
+              })
+            ]);
+          }
+        }
+      }),
+      m.component(Q.components.shared.options.ui.container, {
+        options: [
+          { name: 'onShow', type: 'function', default_value: 'null', description: m.trust('Fired when the tag is clicked.') },
+          { name: 'onRemove', type: 'function', default_value: 'null', description: m.trust('Fired when the remove action is clicked.') },
+        ]
+      })
+    ]);
+  }
+};